chore!: enforce `scram-sha-256` in postgres parameters and `pg_hba.conf` (#122)
## Description

This sets `scram-sha-256` to be the only mechanism that is supported in
the `pg_hba.conf` and the postgres `parameters`.

> [!CAUTION]
> :warning: **BREAKING CHANGE** this change will force `scram-sha-256`
`password_encryption` to better test against a FIPS-valid configuration
for postgres in apps - installations should not be using `md5` since it
is not a valid FIPS 140-3 algorithm but if they are they will need to
rotate the user passwords.

## Related Issue

Fixes #N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-postgres-operator/blob/main/CONTRIBUTING.md#developer-workflow)
followed