fix: add conditional patroni service manifest (#125)
## Description

In sidecar mode, the postgres-operator pod is unable to communicate with
the patroni api on port 8008 to manage updates to the pg-cluster. From
local testing, this does not occur in ambient mode, although I do not
entirely understand why.

- Adds a conditional service manifest to expose port 8008 so the
operator is able to communicate with the cluster as expected.
- Minor docs update to reflect changes made in
https://github.com/defenseunicorns/uds-package-postgres-operator/pull/122

## Related Issue

Fixes #94 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-postgres-operator/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------

Signed-off-by: Samuel Sarnowski <samuel@defenseunicorns.com>